Move components into components folder and update app path
