+++ src/components/Movies/Movies.jsx
@@ -1,19 +1,19 @@
-import React from "react";
-import Movie from "../Movie/Movie";
+import React from 'react'
+import Movie from '../Movie/Movie'
 
 const Movies = ({ movies, handleClick }) => {
   const displayMovies = movies.map((item) => {
-    return <Movie key={item.id} singleMovie={item} />;
-  });
+    return <Movie key={item.id} singleMovie={item} />
+  })
   return (
     <div
       onClick={(event) => {
-        handleClick(event);
+        handleClick(event)
       }}
     >
       {displayMovies}
     </div>
-  );
-};
+  )
+}
 
-export default Movies;
+export default Movies
